Don't make me ignore the gitconfig every `rake install`
This is done by checking the current git configuration and creating a temporary file with those settings, then comparing the in-place config with the temporary one. If those are the same, we know nothing has changed. If they are different, we'll re-create a new file and ask for the settings we need, filling in the current settings as defaults if we have them.

This doesn't solve the problem of the .erb not being updated if you update the ~/.gitconfig, which would be nice to solve but I couldn't figure out a way to do it.